Update patch set 6
Patch Set 6:

> (3 comments)
 > 
 > btw, do you have the page loading time compared to using v1?

no sorry. I was testing this against a version of v2 running off my laptop, with a little bit of data from the tests and a little bit added myself which is why I made the mistake with the change in network structure. (V2 wasnt in prod at time of my testing, but it would be good to be able to test this against it proper so that I do not make a similar mistake).

Patch-set: 6