Update patch set 16
Patch Set 16: Code-Review-1

(3 comments)

I mainly reviewed the distil api part, most of the code logic looks good to me. I think the last thing we should pay attention to is to verify the quotation of all regions (at the end of the month) is almost the same with the final invoice of that month. This is what I am gonna do for cloud_trial_monitor script.

Patch-set: 16
Label: Code-Review=-1